backend/requirements.txt
django==1.11
django-allauth==0.33.0
djangorestframework==3.6.4
numpy==1.13.1
drf-nested-routers==0.90.0